Home Page Shows Popular Movies
Wrapped the Results component in a URL processor component called SearchResults that feeds results into it.

Set the SearchResults components to default to showing popular movies by default.

Renamed the Search component to SearchInput.

Added overloads for the useApi() hook so that it properly types the return object based on the supplied argument.