Movie Details Shows More
Renamed the Details component to MovieDetails and added functionality to retrieve a movie's credits.

Removed the marginBottom value from the Footer component.

Expanded the Movie component to show lists for cast and crew members as well as listing the director under the movie title.

Moved the type definitions for the API to their own file.

Expanded the 'movie' API action to allow for specifying a subset of data from a movie. At the moment it only supports looking up a movie's credits.